pyproject.toml
[project]
name = "codex-supervisor"
version = "0.1.0"
description = "AI Security Testing Supervisor for Codex instances"
dependencies = [
    "pyyaml>=6.0",
    "psutil>=5.9.0",
    "aiofiles>=24.0.0",
    "pytz>=2023.3",
    "python-dotenv>=1.0.0",
    "openai>=2.3.0",
    "tiktoken>=0.12.0",
    "pydantic>=2.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["supervisor"]

[project.scripts]
codex-supervisor = "supervisor.supervisor:cli_main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
